fix: use of `LocalizedStringResource` at build time to load suitable wordings (#1124)
Previously some wordings have been declared as LocalizedStringResource,
but this object should be managed at runtime so as to use the suitable bundles and locales to load the expected translated wordings.
Now, with these values as String and only wrapped in views at runtime, the app can use the suitable translations instead of always the default value.

Signed-off-by: Pierre-Yves Lapersonne <dev@pylapersonne.info>